requirements.txt
requests==2.31.0
python-telegram-bot==21.2
python-dotenv==1.0.1
bip-utils==2.9.3
eth-keys==0.5.1
